removed old engine package and renamed new package to "engine"
