--- src/pypoker/engine/hand_logic/__init__.py
@@ -1,7 +1,0 @@
-"""
-Hand Solver Package
--------------------
-
-Module for determining a players best possible hand.
-This package implements an instance of the base HandSolver class for each type of poker game supported by pypoker.
-"""
